qcacld-3.0: Update ll_stats request pending flag properly
In the current driver implementation, host honors LL_STATS request
even in disconnected state. But in disconnected state, the ll_stats
request pending flag is not updated properly, which results in allowing
multiple LL_STATS requests being sent to FW. So, to avoid this issue,
update the flag properly.

Change-Id: I318455492dc69bd540044b593a405eacdf185ff1
CRs-Fixed: 3339867